+++ package.json
@@ -14,6 +14,7 @@
     "mapbox-gl": "^2.0.0",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
+    "react-router-dom": "^5.2.0",
     "react-dom": "^17.0.1",
     "react-scripts": "4.0.1",
     "react-test-renderer": "^17.0.1",
