+++ src/components/Header/Header.jsx
@@ -2,14 +2,10 @@ import React from "react";
 import "./Header.css";
 import {Logo} from "loft-taxi-mui-theme";
 import {Button, Container, Paper} from "@material-ui/core";
-import PropTypes from 'prop-types';
-import { withAuth } from "../../helpers/AuthContext";
+import {withAuth} from "../../helpers/AuthContext";
+import {Link} from "react-router-dom";
 
-const Header = ({customNavigation, logOut}) => {
-    const exit = () => {
-        customNavigation('login');
-        logOut();
-    }
+const Header = ({logOut}) => {
     const buttonList = [
         {
             name: "map",
@@ -31,11 +27,15 @@ const Header = ({customNavigation, logOut}) => {
                         <ul className="navList">
                             {buttonList.map(({name, text}) => (
                                 <li className="navItem" key={name}>
-                                    <Button data-testid={'nav_'+name} onClick={() => customNavigation(name)}>{text}</Button>
+                                    <Button data-testid={'nav_' + name}>
+                                        <Link to={name}>{text}</Link>
+                                    </Button>
                                 </li>
                             ))}
                             <li className="navItem">
-                                <Button data-testid="nav_exit" onClick={() => exit()}>Выйти</Button>
+                                <Link to='logIn'>
+                                    <Button data-testid="nav_exit" onClick={() => logOut()}>Выйти</Button>
+                                </Link>
                             </li>
                         </ul>
                     </nav>
@@ -43,9 +43,6 @@ const Header = ({customNavigation, logOut}) => {
             </Container>
         </Paper>
     );
-}
-Header.prototype = {
-    customNavigation: PropTypes.func.isRequired
 }
 
 export default withAuth(Header);
